Compare clean terminator frame widths explicitly
Parity trace terminators store final_frame as u64 while the engine counter remains u32. Widen the engine value for the strict clean-terminator assertion so cache-v12 example builds retain exact comparison without truncation.

+++ crates/robin_rs/examples/original_parity_replay.rs
@@ -2544,7 +2544,7 @@ fn run_replay(options: Options, visual_window: Option<robin_rs::window::GameWind
                 "parity terminator final_frame disagrees with its initial frame and frame count"
             );
             assert_eq!(
-                engine.frame_counter(),
+                u64::from(engine.frame_counter()),
                 final_frame,
                 "Rust final frame disagrees with the clean parity terminator"
             );
